Add codeset to locale on non-Windows systems
The call to setlocale() should contain the current codeset (e.g. en_US.UTF-8) instead
of just the language and territory (en_US). Otherwise the user would have to have
loacles generated for "en_US" as well which is often not the case with newer Linux
distributions (e.g. Ubuntu 8.10). Fix for bug 338.