Enable timecode generation for appended AVC packetizers
When a packetizer is appended to another one then the ::set_headers() member
function of the appended one will not be called. The AVC ES packetizer enables
the timecode generation of the parser in ::set_headers(). So the timecode
generation is always off in the appended packetizers.

This is a problem when the user appends raw AVC ES files because for them
the timecode generation must turned on because otherwise there will be no
timecodes for the frames. So now the AVC ES packetizer hoooks into the
::connect() function and enables timecode generation if neccessary.

Possible fix for bug 344.