Use GetOEMCP() for getting the local code page instead of GetACP().
Windows uses two code pages; one for cmd.exe and one for the rest.
They can be different (e.g. on a German Windows). As most programs
output text to the console unless the user overrides this with
--redirect-output the programs should use cmd.exe's code page
as their default code page. This is what GetOEMCP() does.

Overriding it with --output-charset is still possible.