Replaced the wxTreebook in the header editor with a wxTreeCtrl and my own control logic.
The reason is that the wxTreebook offers the program almost no control over
the size of the tree and the pages. This causes weird effects both on Linux
and on Windows: the tree is often way to small, it gets resized each time
the user selects an item, tree items are not redrawn properly on Windows.