Added a hack for adding "stop displaying" commands to certain VobSub subtitle packets.
Patch by Todd Schmuland.

Explanation by Todd Schmuland:

I own a WDTV.  So far it's the only hardware media player that can
handle VobSub subtitles.  It's not perfect however.  One major flaw it
has is not displaying subtitles that have no SPU duration embedded in
them (the cause of the vobsub_reader: Could not extract the duration
for a SPU packet).  The normal behavior a player should do is display
the subtitle frame until the next subtitle is displayed, but the WDTV
simply chooses to not display the subtitle at all.

So I wrote a hack --engage vobsub_subpic_stop_cmds that writes in a
Stop Display command into these packets.  The end result is that the
WDTV displays the subtitles now.  There is a slight flicker if the two
subtitle frames are the same, but it's a lot better than not seeing
the subtitle at all.  Since this is adding SPU content not in the
original VobSub, I used the hacks mechanism to implement it.
Hopefully WD will properly handle VobSub subtitle frames with no Stop
Display command in a future firmware update, but until then I'll use
this hack.