Improved detection of wrong padding packet lengths in MPEG program streams.
Fix for bug 393.

Explanation by Todd Schmuland:

I looked at his mpeg file and my assumption of packets always being on
a page boundary was incorrect.  His mpeg isn't perfect (each specified
audio packet length is 16 bytes short of what it really is, hence all
the lost sync messages), but the parser is capable of dealing with it.
So I have to remove the erroneous page boundary limitation.

Attached is the fixed r_mpeg_ps.cpp.  The new behavior now is thus.
Whenever a pad packet is encountered, the code now skips forward the
packet length specified and checks to see if the next header is there.
If it is then great, if not then it searches for the next header and
calculates the correct pad packet length.  This removes the page
boundary limitation.  I tested it on Transformers and it works well.
I also tested it against his mpeg and it works as well.