Corrected the "operator <" logic for file_id_t structures.
The device number has precendence over the inode number. Therefore the
inode numbers must only be compared if the device numbers are equal.

This resulted in some strange cases in which the std::map<> object for
which file_id_t structures were used as keys was removing/inserting
the wrong elements. This caused crashes in rare cases, e.g. for access
to files on VFAT formatted file systems because the device numbers
were higher than ones for ext3 or devfs (think "/dev/urandom").