Refined the I frame detection for AVC/h.264 video tracks.
The decision whether or not to merge two NALUs into a single I frame
needed to take the "first macroblock in frame" counter into
account. Otherwise a single I frame might be detected as being two or
more separate frames if one of the other conditions for NALU
separation was met. This resulted in garbled playback for the current
group of pictures as well as in wrong timestamps being issued to
following frames.

Fix for bug 415.