If the AVC packetizer must not create timecodes then the demuxer must provide them.
The AVC OGM demuxer is now derived from the general MS compatibility
video OGM demuxer because that one already calculates the timecodes
for each packet and passes them on to the packetizer. The AVC
packetizer needs either timecodes provided by the demuxer or it must
be allowed to generate its own.

Fix for bug 418.