Always quote the mkvmerge executable path and name.
Windows' execute function works OK if the path to an .exe is unquoted
and contains a single space, probably to account for "C:\Program
Files\...". However, if the path contains two spaces then the
application to execute will not be found crashing mmg which did not
anticipate that mkvmerge cannot be spawned.

Fix for bug 419.