Fix DisplayWidth/Height values that are used for aspect ratio storage only
Some muxers (x264) only use the quotient of DisplayWidth/Height for
storing the aspect ratio and make the actual values very
small (e.g. 16x9 or 4x3). This is bad for demuxers that actually
follow the spec and use these values as the intended display
width/height in pixels.

So re-calculate sensible values for DisplayWidht & DisplayHeight based
on the values of PixelWidth & PixelHeight and the quotient
DisplayWidth / DisplayHeight.