Always make sure memory is free after a packet has been added
It's possible that even though a compressor is set the compressor
itself will not modify the data (for an instance of
compressor_c()). In such a case the data block might not be free,
e.g. if it is read from a Matroska file and not modified by the
packetizer. So always grab() the data.