Always flush all packetizers
If --default-duration is used then the timecode factory applies
queueing. If the packetizer is not flushed then the timecode factory
is not flushed either. If the video frame does not end in a key frame
then the timecode factory will have frames left, and the packet queue
won't be empty resulting in lost frames.