Fix compilation with mingw
If wxWidgets is built as a DLL then the DECLARE_EVENT_TYPE macro
assumes the event itself is defined in a DLL. If it isn't then the
linker will look for the wrong symbol.

Therefore don't use the macro as the event does indeed not live in a
DLL in this case.