Make mkvinfo's verbosity level independent from the verbosity level of the common code
Code in src/common uses the global "verbose" variable for debugging
output. mkvinfo used the same variable for determining how much data
to output. This clashed if output was redirected and verbose was >= 2
as the mm_write_cache_io_c class that's used for redirecting the
output was outputting a lot of info in verbose mode 2 ultimately
causing mkvinfo to crash.