Don't use slow UpdateSize() when splitting is active
Fix for bug 611.

When splitting is active and each video frame is a key frame (happens
e.g. with files created by MakeMKV) then each video frame added to the
queue caused an update for the cues size. This is rather
slow. Therefore simply estimate the size of a cue entry with 18
bytes (it's 16 bytes for files approximately 20 MB big, 18 bytes for
files 1 GB big).