Fix timecodes in appended subtitle tracks from compound containers
If subtitle tracks are appended to video tracks from the same file
(e.g. two Matroska files) then the subtitle timecodes of the second
and all following subtitle files were based on the last timecode in
the first subtitle track instead of the whoe file's last timecode. Fix
for bug 620.

This bug was introduced in 06fa821.