Build system: Always process source files in alphabetical order
Background is that the po files contain the file names in the order
they're given to xgettext. Unfortunately that order depends on the
order in which Ruby (via Rake) finds the files in the file
system. That again depends on the OS and file system involved.

All this leads to the situation that, even though no translation or
translatable string has changed, almost all of a po file changes just
because the file names are ordered differently.

This can be avoided by always handing xgettext a list of file names
that's already sorted.