Matroska identify verbose: Add more properties
default_duration is added for all track types if the element exists.

audio_sampling_frequency and audio_channels are output for audio
tracks. As audio_sampling_frequency has a default value in the
Matroska specification (8000 Hz) it will always be output even if the
header element is not present in the file. audio_channels will only be
output if the header element is present in the file.