Common: fix building with clang++
clang(++) doesn't support the type __float128. gcc's libstdc++ uses
that type but only if __STRICT_ANSI__ is not defined.

clang++ -std=c++11 does set __STRICT_ANSI__, but common.h #undefines
it due to other compilation issues with mingw in the past.

So include <type_traits> (in which __float128 occurs) before
#undefining __STRICT_ANSI__.